pyproject.toml
[project]
name = "remind-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amqp==5.3.1",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "billiard==4.2.1",
    "celery==5.5.3",
    "certifi==2025.6.15",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "decouple>=0.0.7",
    "django==5.2.3",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "kombu==5.5.4",
    "packaging==25.0",
    "pip==25.1.1",
    "prompt-toolkit==3.0.51",
    "pyjwt==2.9.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-telegram-bot==22.1",
    "redis==6.2.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
    "vine==5.1.0",
    "wcwidth==0.2.13",
]
